Updated locations of code
